Update Calculator.java
Убрал лишнее приведение

+++ src/com/startjava/lesson_2_3_4/calculator/Calculator.java
@@ -28,7 +28,7 @@ public class Calculator {
                 result = Math.pow(a, b);
                 break;
             case '%':
-                result = (double) a % b;
+                result = a % b;
                 break;
         }
         if (result % 1 == 0) {
